+++ engines/scumm/saveload.cpp
@@ -1242,7 +1242,7 @@ void ScummEngine::saveOrLoad(Serializer *s) {
 			}
 			s->saveUint16(0xFFFF);	// End marker
 		} else {
-			while ((type = (ResType)s->loadUint16()) != 0xFFFF) {
+			while ((int)(type = (ResType)s->loadUint16()) != 0xFFFF) {
 				while ((idx = s->loadUint16()) != 0xFFFF) {
 					assert(idx < _res->_types[type].size());
 					loadResource(s, type, idx);
@@ -1430,7 +1430,7 @@ void ScummEngine::saveOrLoad(Serializer *s) {
 			}
 		s->saveByte(0xFF);
 	} else {
-		while ((type = (ResType)s->loadByte()) != 0xFF) {
+		while ((int)(type = (ResType)s->loadByte()) != 0xFF) {
 			idx = s->loadUint16();
 			_res->lock(type, idx);
 		}
